[Hashmap] Use Array of Lists instead of List**
It is easier to follow the code.
I make use as much as possible of other data structures.
It is an example of using a data structure inside another one.
Raw pointers arrays (e.g.: void**) are used in the B-Trees data
structures, because they use the system call memmove to rebalance the
tree. That way, we can relocate all data from one node in one call,
instead of looping through all the keys and pointers of a node.